Remove the sysdep interface since it is fixed at compilation time
It should be enough to specialize sysdep_posix.c if another platform need
to be supported and link it. The posix functions does not depend on the
image format to handle. They are dependent on the plateform on which the
code is compiled.

This make the code a lot easier to maintain since we do not have to go
through function's pointers that are fixed at compilation time.

Signed-off-by: Patryck Rouleau <pfrouleau@gmail.com>