fix: partclone's v2 image always has a CRC after the bitmap
The option "-a" allow the user to disable the checksum for the data blocks
and it can also allow the user to use a different algorithm than CRC32,
however this does not affect the presence of a CRC32 for the bitmap. This
checksum is important to detect corruption in the bitmap, which would
make the blocks unusable.

Signed-off-by: Patrick Rouleau <pfrouleau@gmail.com>