WIP: libpartclone: read image 0002's bitmap CRC
Reading the CRC improve partclone_imageinfo output, in particular for
images made with "-a0"; they do not show "4 byte trailer" anymore.
This also align the v2 code with the v1 code, which reads the 0001's
bitmap signature (ie "BiTmAgIc").

TODO: double-check if that cause issue for imagemount.

The sizeof(unsigned char) are removed:
  - to make the lines shorter,
  - because they are not needed (==1);
  - because one was missing and removing the other ones make the code
    uniform.