WIP: imagemount: move request state variables into a context struct
This is a preliminary step in preparation to split nbd_service_requests()
into many targeted functions.

Note: It is not clear why the callback functions are working with pointers
instead of "static global" variables. The pointers are still global...
The pointers were not declared volatile, while the pointees were. The
volatile is required and maybe the pointers have another side-effect.
The pointers are kept until this is clarified. The context struct may be
a little bit overkill... we will see how it turns out with the smaller
functions.