WIP: imagemount: move the sigaction into req_context
Since the code depending on sigaction variables will be moved in distinct
methods, they need to be moved into the context to be accessible by these
future methods.

While at it, lets encapsulate the detail of installing the reapchild
callback function.

TODO: Should we encapsulate the parameters too and simply name the function
set_child_supervisor() or set_child_caretaker()? To the extrem, even the
fork() could be encapsulated into that function...