imagemount: move a big chunk of code outside of nbd_service_requests()
This make the nbd_service_requests's while-loop manageable and all the
code handling the request is now in handle_request(). This also allows
to reduce the scope of a few variables.

NB: since this commit changes the indentation of many lines, "git diff"
shows different changes with or without "-w".