WIP: create handle_*() functions to clarify the code
These changes need to be done in a different way to reduce the amount of
lines modified at once.

By example:
- create send_req_reply()
- create handle_read_request() + send_read_reply()
- create handle_write_request()

And maybe the if/else inversion should be done in their own commits...

TODO: Why do we get a log message like this? It was already there before
this commit.

    kernel: [20124.148132] block nbd1: Send disconnect failed -32