imagemount: return a simulated NTFS's backup boot sector if it is missing
partclone does not include the NTFS's backup boot sector in the image
because NTFS itself exclude it from its block count. However, recent
mount's version and/or NTFS driver refuse to mount the image when the
backup sector is missing.

The backup boot sector is simulated by using a copy of the boot sector.

We need to allocate one extra byte for the "byte"map to avoid
v1_blockused() to access unallocated memory.